pyproject.toml
[project]
name = "my-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "frontend>=0.0.3",
    "google-api-python-client>=2.169.0",
    "google-auth-oauthlib>=1.2.2",
    "loguru>=0.7.3",
    "openai>=1.77.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
]
